Crate gear_backend_common

Crate provides support for wasm runtime.

Re-exports

pub use log;

Modules

funcs

Syscall implementations generic over wasmi and sandbox backends.

lazy_pages

Core logic for usage both in runtime and in lazy-pages native part.

memory

Work with WASM program memory in backends.

runtime

Trait that both sandbox and wasmi runtimes must implement.

Macros

assert_err

assert_ok

syscall_args_trace

syscall_trace

Structs

BackendReport

ExtInfo

LimitedStr

Wrapped string to fit core_backend::TRIMMED_MAX_LEN amount of bytes.

SystemReservationContext

SystemTerminationReason

Non-actor related termination reason.

Enums

ActorTerminationReason

EnvironmentError

TerminationReason

TrapExplanation

UnrecoverableExecutionError

Execution error in infallible sys-call.

UnrecoverableExtError

UnrecoverableMemoryError

Memory error in infallible sys-call.

UnrecoverableWaitError

Wait error in infallible sys-call.

Constants

PTR_SPECIAL

Traits

BackendAllocSyscallError

A trait for conversion of the externalities memory management error to api error.

BackendExternalities

Extended externalities that can manage gas counters.

BackendState

BackendSyscallError

A trait for conversion of the externalities API error to TerminationReason and RunFallibleError.

BackendTermination

A trait for termination of the gear sys-calls execution backend.

Environment

Type Definitions

EnvironmentExecutionResult