Module gear_backend_common::lazy_pages

Core logic for usage both in runtime and in lazy-pages native part.

Structs

GlobalsAccessError

Globals access error.

GlobalsConfig

Globals ctx for lazy-pages initialization for program.

LazyPagesWeights

Lazy-pages cases weights.

Enums

GlobalsAccessMod

Informs lazy-pages whether they work with native or WASM runtime.

Status

Lazy-pages status. By default in program initialization status is set as Normal. If nothing bad happens in lazy-pages, then status remains to be Normal. If gas limit exceed, then status is set as GasLimitExceeded, and lazy-pages starts to skips all signals processing until the end of execution. The same is for gas allowance exceed, except it sets status as GasAllowanceExceed. In the end of execution this status is checked and if it’s not Normal then termination reason sets as gas limit exceeded or gas allowance exceeded, depending on status. NOTE: repr(i64) is important to be able add additional fields, without old runtimes separate support logic.

Traits

GlobalsAccessor

Globals access trait.